requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.22.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
attrs==25.4.0
black==26.1.0
certifi==2026.1.4
cffi==2.0.0
click==8.3.1
fastapi==0.128.0
frozenlist==1.8.0
greenlet==3.3.1
h11==0.16.0
idna==3.11
multidict==6.7.1
mypy_extensions==1.1.0
packaging==26.0
pathspec==1.0.4
platformdirs==4.5.1
propcache==0.4.1
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
PyJWT==2.11.0
python-dotenv==1.2.1
pytokens==0.4.1
SQLAlchemy==2.0.46
starlette==0.50.0
tenacity==9.1.2
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
yarl==1.22.0
